Version 1.0
Created the first version of the backend. This backend is ready to register and login users, stream videos and upload them, manage them within the backend, etc. Some things need to be added in future commits. For example: User passwords are not beeing hashed currently before beeing saved in the database. Also, necessary paths are currently managed within the CMakeLists.txt. I want to change this to a configurable ini file in later updates. Currently building and hosting was only tested using windows 10 64 bit/compiled with vs2022 Community